requirement.txt
python-bitcoinlib
python-bitcoinrpc